OPS: Rename CLI and add it as a script in pyproject.toml
