+++ .gitignore
@@ -1,1 +1,2 @@
 .idea/
+__pycache__
