+++ pyproject.toml
@@ -6,10 +6,12 @@ authors = ["Marcus Lugg <marcus@octue.com>"]
 readme = "README.md"
 packages = [{include = "generate_pull_request_description"}]
 
+[tool.poetry.scripts]
+generate-pull-request-description = "generate_pull_request_description.generate_pull_request_description:main"
+
 [tool.poetry.dependencies]
 python = "^3.10"
 requests = "^2.28.1"
-
 
 [tool.poetry.group.dev.dependencies]
 coverage = "^6.5.0"
